+++ lib/http.go
@@ -399,7 +399,11 @@ func (s *Server) stripBasePrefixFromRequest(r *http.Request) *http.Request {
 }
 
 func (s *Server) ServeHTTPNext(w http.ResponseWriter, r *http.Request) {
-	if target, ok := s.opts.TargetRoutes[r.Host]; ok {
+	routes := s.opts.TargetRoutes
+	if s.routesWatcher != nil {
+		routes = s.routesWatcher.Get()
+	}
+	if target, ok := routes[r.Host]; ok {
 		rp, err := s.getReverseProxyForTarget(target)
 		if err != nil {
 			lg, _ := s.getRequestLogger(r)